Changed a typedef int to typedef long to make this module more 64-bit friendly (the ints are being stored on a Stack of void pointers, and the size of these pointers changes depending on whether the target platform is 32- or 64-bit.  For gcc on Linux and MacOS X:
  - 32-bit platform:
      - sizeof(int)  = 4
      - sizeof(long) = 4

  - 64-bit platform:
      - sizeof(int)  = 8
      - sizeof(long) = 8

+++ src/sbml/MathMLTagCodes.hpp
@@ -65,7 +65,7 @@
 #endif  // USE_EXPAT
 
 
-typedef int MathMLTagCode_t;
+typedef long MathMLTagCode_t;
 
 static const MathMLTagCode_t MATHML_TAG_ABS            =  0;
 static const MathMLTagCode_t MATHML_TAG_AND            =  1;
